[refactor] : WeatherCardSection.jsx 위치 변경
